+++ crates/da-clients/ethereum/src/lib.rs
@@ -20,8 +20,8 @@ pub struct EthereumDaClient {
 #[async_trait]
 impl DaClient for EthereumDaClient {
     async fn publish_state_diff(&self, _state_diff: Vec<Vec<u8>>, _to: &[u8; 32]) -> Result<String> {
-        // Here in case of ethereum we are not publishing the state diff because we are doing it all together in update_state job.
-        // So we don't need to send the blob here.
+        // Here in case of ethereum we are not publishing the state diff because we are doing it all
+        // together in update_state job. So we don't need to send the blob here.
         Ok("NA".to_string())
     }