+++ crates/orchestrator/src/data_storage/aws_s3/mod.rs
@@ -1,11 +1,12 @@
-use crate::data_storage::aws_s3::config::AWSS3Config;
-use crate::data_storage::DataStorage;
 use async_trait::async_trait;
 use aws_sdk_s3::config::{Builder, Credentials, Region};
 use aws_sdk_s3::primitives::ByteStream;
 use aws_sdk_s3::Client;
 use bytes::Bytes;
 use color_eyre::Result;
+
+use crate::data_storage::aws_s3::config::AWSS3Config;
+use crate::data_storage::DataStorage;
 
 /// Module for AWS S3 config structs and implementations
 pub mod config;