+++ crates/orchestrator/src/jobs/mod.rs
@@ -4,24 +4,30 @@ use std::time::Duration;
 use async_trait::async_trait;
 use color_eyre::eyre::eyre;
 use color_eyre::Result;
-use mockall::{automock, mock};
+use mockall::automock;
+use mockall_double::double;
 use tracing::log;
 use uuid::Uuid;
 
 use crate::config::{config, Config};
 use crate::jobs::constants::{JOB_PROCESS_ATTEMPT_METADATA_KEY, JOB_VERIFICATION_ATTEMPT_METADATA_KEY};
+#[double]
+use crate::jobs::job_handler_factory::factory;
 use crate::jobs::types::{JobItem, JobStatus, JobType, JobVerificationStatus};
 use crate::queue::job_queue::{add_job_to_process_queue, add_job_to_verification_queue};
 
 pub mod constants;
 pub mod da_job;
+pub mod job_handler_factory;
 pub mod proving_job;
 pub mod register_proof_job;
 pub mod snos_job;
 pub mod state_update_job;
+
 /// The Job trait is used to define the methods that a job
 /// should implement to be used as a job for the orchestrator. The orchestrator automatically
 /// handles queueing and processing of jobs as long as they implement the trait.
+#[automock]
 #[async_trait]
 pub trait Job: Send + Sync {
     /// Should build a new job item and return it
@@ -49,28 +55,6 @@ pub trait Job: Send + Sync {
     fn verification_polling_delay_seconds(&self) -> u64;
 }
 
-mock! {
-    pub Job {}
-    impl Clone for Job {
-        fn clone(&self) -> Self;
-    }
-
-    #[async_trait]
-    impl Job for Job {
-        async fn create_job(
-            &self,
-            config: &Config,
-            internal_id: String,
-            metadata: HashMap<String, String>,
-        ) -> Result<JobItem>;
-        async fn process_job(&self, config: &Config, job: &mut JobItem) -> Result<String>;
-        async fn verify_job(&self, config: &Config, job: &mut JobItem) -> Result<JobVerificationStatus>;
-        fn max_process_attempts(&self) -> u64;
-        fn max_verification_attempts(&self) -> u64;
-        fn verification_polling_delay_seconds(&self) -> u64;
-    }
-}
-
 pub mod types;
 
 /// Creates the job in the DB in the created state and adds it to the process queue
@@ -86,9 +70,8 @@ pub async fn create_job(job_type: JobType, internal_id: String, metadata: HashMa
         ));
     }
 
-    let job_handler = get_job_handler(&job_type).await;
+    let job_handler = factory::get_job_handler(&job_type).await;
     let job_item = job_handler.create_job(config.as_ref(), internal_id, metadata).await?;
-    println!(">>>> job_item : {:?}", job_item);
     config.database().create_job(job_item.clone()).await?;
 
     add_job_to_process_queue(job_item.id).await?;
@@ -117,7 +100,7 @@ pub async fn process_job(id: Uuid) -> Result<()> {
     // outdated
     config.database().update_job_status(&job, JobStatus::LockedForProcessing).await?;
 
-    let job_handler = get_job_handler(&job.job_type).await;
+    let job_handler = factory::get_job_handler(&job.job_type).await;
     let external_id = job_handler.process_job(config.as_ref(), &mut job).await?;
     let metadata = increment_key_in_metadata(&job.metadata, JOB_PROCESS_ATTEMPT_METADATA_KEY)?;
 
@@ -151,7 +134,7 @@ pub async fn verify_job(id: Uuid) -> Result<()> {
         }
     }
 
-    let job_handler = get_job_handler(&job.job_type).await;
+    let job_handler = factory::get_job_handler(&job.job_type).await;
     let verification_status = job_handler.verify_job(config.as_ref(), &mut job).await?;
 
     match verification_status {
@@ -195,22 +178,6 @@ pub async fn verify_job(id: Uuid) -> Result<()> {
     };
 
     Ok(())
-}
-
-#[cfg(not(test))]
-async fn get_job_handler(job_type: &JobType) -> Box<dyn Job> {
-    match job_type {
-        JobType::DataSubmission => Box::new(da_job::DaJob),
-        JobType::SnosRun => Box::new(snos_job::SnosJob),
-        JobType::ProofCreation => Box::new(proving_job::ProvingJob),
-        JobType::StateTransition => Box::new(state_update_job::StateUpdateJob),
-        _ => unimplemented!("Job type not implemented yet."),
-    }
-}
-
-#[cfg(test)]
-async fn get_job_handler<'a>(_job_type: &JobType) -> Box<dyn Job> {
-    return Box::new(config().await.job_handler.clone());
 }
 
 async fn get_job(id: Uuid) -> Result<JobItem> {