+++ crates/orchestrator/src/jobs/state_update_job/mod.rs
@@ -7,16 +7,14 @@ use async_trait::async_trait;
 use cairo_vm::Felt252;
 use color_eyre::eyre::eyre;
 use color_eyre::Result;
+use settlement_client_interface::SettlementVerificationStatus;
 use snos::io::output::StarknetOsOutput;
 use uuid::Uuid;
-
-use settlement_client_interface::SettlementVerificationStatus;
 
 use super::constants::{
     JOB_METADATA_STATE_UPDATE_ATTEMPT_PREFIX, JOB_METADATA_STATE_UPDATE_LAST_FAILED_BLOCK_NO,
     JOB_PROCESS_ATTEMPT_METADATA_KEY,
 };
-
 use crate::config::{config, Config};
 use crate::constants::SNOS_OUTPUT_FILE_NAME;
 use crate::jobs::constants::JOB_METADATA_STATE_UPDATE_BLOCKS_TO_SETTLE_KEY;
@@ -76,13 +74,20 @@ impl Job for StateUpdateJob {
         self.insert_attempts_into_metadata(job, &attempt_no, &sent_tx_hashes);
 
         // external_id returned corresponds to the last block number settled
-        Ok(block_numbers.last().expect("Last number in block_numbers array returned as None. Possible Error : Delay in job processing or Failed job execution.").to_string())
+        Ok(block_numbers
+            .last()
+            .expect(
+                "Last number in block_numbers array returned as None. Possible Error : Delay in job processing or \
+                 Failed job execution.",
+            )
+            .to_string())
     }
 
     /// Returns the status of the passed job.
     /// Status will be verified if:
     /// 1. the last settlement tx hash is successful,
-    /// 2. the expected last settled block from our configuration is indeed the one found in the provider.
+    /// 2. the expected last settled block from our configuration is indeed the one found in the
+    ///    provider.
     async fn verify_job(&self, config: &Config, job: &mut JobItem) -> Result<JobVerificationStatus> {
         let attempt_no =
             job.metadata.get(JOB_PROCESS_ATTEMPT_METADATA_KEY).expect("Could not find current attempt number.").clone();
@@ -115,7 +120,7 @@ impl Job for StateUpdateJob {
                             return Ok(new_status.into());
                         }
                         SettlementVerificationStatus::Pending => {
-                            return Err(eyre!("Tx {tx_hash} should not be pending."))
+                            return Err(eyre!("Tx {tx_hash} should not be pending."));
                         }
                         SettlementVerificationStatus::Verified => {}
                     }