+++ crates/orchestrator/src/tests/common/mod.rs
@@ -22,7 +22,6 @@ use crate::database::{DatabaseConfig, MockDatabase};
 use crate::jobs::types::JobStatus::Created;
 use crate::jobs::types::JobType::DataSubmission;
 use crate::jobs::types::{ExternalId, JobItem};
-use crate::jobs::MockJob;
 use crate::queue::MockQueueProvider;
 
 pub async fn init_config(
@@ -55,7 +54,6 @@ pub async fn init_config(
         Box::new(database),
         Box::new(queue),
         Box::new(storage_client),
-        MockJob::new(),
     )
 }