+++ crates/orchestrator/src/tests/data_storage/mod.rs
@@ -1,11 +1,12 @@
-use crate::data_storage::aws_s3::config::AWSS3Config;
-use crate::data_storage::aws_s3::AWSS3;
-use crate::data_storage::{DataStorage, DataStorageConfig};
-use crate::tests::config::TestConfigBuilder;
 use bytes::Bytes;
 use rstest::rstest;
 use serde_json::json;
 use utils::env_utils::get_env_var_or_panic;
+
+use crate::data_storage::aws_s3::config::AWSS3Config;
+use crate::data_storage::aws_s3::AWSS3;
+use crate::data_storage::{DataStorage, DataStorageConfig};
+use crate::tests::config::TestConfigBuilder;
 
 /// This test checks the ability to put and get data from AWS S3 using `AWSS3`.
 /// It puts JSON data into a test bucket and retrieves it, verifying the data