+++ crates/orchestrator/src/tests/database/mod.rs
@@ -1,9 +1,10 @@
+use rstest::*;
+use uuid::Uuid;
+
 use crate::config::config;
 use crate::jobs::types::{ExternalId, JobItem, JobStatus, JobType};
 use crate::tests::common::drop_database;
 use crate::tests::config::TestConfigBuilder;
-use rstest::*;
-use uuid::Uuid;
 
 #[rstest]
 #[tokio::test]